pyproject.toml
[project]
name = "rezeror"
version = "0.1.0"
description = "Re:Zero web novel parser + Flask reader"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
	"beautifulsoup4>=4.13.0",
	"flask>=3.1.0",
	"markdown>=3.8",
	"markdownify>=1.1.0",
	"pyyaml>=6.0.2",
	"requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
	"pytest>=8.4.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]

[project.scripts]
rezeror = "rezeror.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rezeror"]
